Share a neutral shell across built-in and plugin interaction forms
