+++ apps/app/src/components/plugin/PluginPendingInteractionComposer.test.tsx
@@ -94,21 +94,74 @@ describe("PluginPendingInteractionComposer", () => {
     });
     fireEvent.keyDown(window, { key: "1" });
     expect(onShortcut).toHaveBeenCalledTimes(1);
-    const toggle = screen.getByRole("button", { name: "Collapse form" });
+    const toggle = screen.getByRole("button", { name: "Hide details" });
     toggle.focus();
     fireEvent.click(toggle);
     expect(screen.queryByRole("textbox")).toBeNull();
     expect(document.activeElement).toBe(
-      screen.getByRole("button", { name: "Expand form" }),
+      screen.getByRole("button", { name: "Show details" }),
     );
     fireEvent.keyDown(window, { key: "2" });
     expect(onShortcut).toHaveBeenCalledTimes(1);
-    fireEvent.click(screen.getByRole("button", { name: "Expand form" }));
+    fireEvent.click(screen.getByRole("button", { name: "Show details" }));
     expect(screen.getByRole("textbox").getAttribute("value")).toBe(
       "Keep my draft",
     );
     fireEvent.keyDown(window, { key: "3" });
     expect(onShortcut).toHaveBeenCalledTimes(2);
+    fireEvent.keyDown(screen.getByRole("textbox"), { key: "Escape" });
+    expect(screen.queryByRole("textbox")).toBeNull();
+    expect(document.activeElement).toBe(
+      screen.getByRole("button", { name: "Show details" }),
+    );
+    fireEvent.click(screen.getByRole("button", { name: "Show details" }));
+    expect(screen.getByRole("textbox").getAttribute("value")).toBe(
+      "Keep my draft",
+    );
+  });
+
+  it("opens a new interaction with a fresh form after the previous one was collapsed", () => {
+    function Renderer() {
+      const [answer, setAnswer] = useState("");
+      return (
+        <input
+          aria-label="Answer"
+          value={answer}
+          onChange={(event) => setAnswer(event.target.value)}
+        />
+      );
+    }
+    setPluginSlotRegistrations(
+      "secrets",
+      registrations([{ id: "secret-request", component: Renderer }]),
+    );
+    const client = new QueryClient();
+    const composer = (id: string) => (
+      <QueryClientProvider client={client}>
+        <PluginPendingInteractionComposer
+          interaction={{ ...interaction, id }}
+          request={{
+            pluginId: "secrets",
+            rendererId: "secret-request",
+            title: interaction.payload.title,
+            data: interaction.payload.data,
+          }}
+          dismissal="cancel"
+        />
+      </QueryClientProvider>
+    );
+    const view = render(composer(interaction.id));
+    fireEvent.change(screen.getByRole("textbox"), {
+      target: { value: "Previous answer" },
+    });
+    fireEvent.click(screen.getByRole("button", { name: "Hide details" }));
+    view.rerender(composer("pint_new"));
+    expect(screen.getByRole("textbox").getAttribute("value")).toBe("");
+    expect(
+      screen
+        .getByRole("button", { name: "Hide details" })
+        .getAttribute("aria-expanded"),
+    ).toBe("true");
   });
 
   it("mounts only the renderer registered by the interaction's plugin", () => {
