+++ apps/app/src/components/plugin/PluginPendingInteractionComposer.tsx
@@ -1,5 +1,8 @@
-import { Activity, useCallback, useId, useMemo, useState } from "react";
-import { Icon } from "@bb/shared-ui/icon";
+import {
+  PendingInteractionShell,
+  type PendingInteractionSourceThread,
+} from "@/components/thread/pending-interactions/PendingInteractionShell";
+import { useCallback, useMemo, useState } from "react";
 import { Button } from "@bb/shared-ui/button";
 import type { JsonValue, PendingInteraction } from "@bb/domain";
 import { PluginSlotMount } from "./PluginSlotMount";
@@ -22,22 +25,19 @@ interface PluginPendingInteractionComposerProps {
   >;
   request: PluginPendingInteractionRequest;
   dismissal: "cancel" | "stop-turn";
+  sourceThread?: PendingInteractionSourceThread;
 }
 
 export function PluginPendingInteractionComposer({
   interaction,
   request,
   dismissal,
+  sourceThread,
 }: PluginPendingInteractionComposerProps) {
   const { pendingInteractions } = usePluginSlots();
   const stopThread = useStopThread();
   const [error, setError] = useState<string | null>(null);
   const [submitting, setSubmitting] = useState(false);
-  const [collapsedInteractionId, setCollapsedInteractionId] = useState<
-    string | null
-  >(null);
-  const isCollapsed = collapsedInteractionId === interaction.id;
-  const contentId = useId();
   const slot = useMemo(
     () =>
       resolvePendingInteraction(
@@ -89,42 +89,16 @@ export function PluginPendingInteractionComposer({
   const dismissLabel = dismissal === "cancel" ? "Cancel" : "Stop turn";
 
   return (
-    <section className="mb-2 rounded-lg border border-border bg-surface-recessed text-xs text-muted-foreground">
-      <header className="flex min-w-0 items-center gap-2 py-1.5 pl-4 pr-1.5">
-        <button
-          type="button"
-          aria-expanded={!isCollapsed}
-          aria-controls={contentId}
-          onClick={() =>
-            setCollapsedInteractionId(isCollapsed ? null : interaction.id)
-          }
-          className="min-w-0 flex-1 rounded-md py-1 text-left focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring"
-          title={request.title}
-        >
-          <h3 className="truncate text-sm font-semibold text-foreground">
-            {request.title}
-          </h3>
-        </button>
-        <Button
-          type="button"
-          size="icon"
-          variant="ghost"
-          aria-label={isCollapsed ? "Expand form" : "Collapse form"}
-          aria-expanded={!isCollapsed}
-          aria-controls={contentId}
-          onClick={() =>
-            setCollapsedInteractionId(isCollapsed ? null : interaction.id)
-          }
-          className="size-7 shrink-0 text-subtle-foreground/75"
-        >
-          <Icon
-            name={isCollapsed ? "ChevronUp" : "ChevronDown"}
-            className="size-3.5"
-          />
-        </Button>
-      </header>
-      <Activity mode={isCollapsed ? "hidden" : "visible"}>
-        <div id={contentId} className="px-4 pb-3">
+    <PendingInteractionShell
+      key={interaction.id}
+      label={request.title}
+      initiallyExpanded
+      errorMessage={error}
+      sourceThread={sourceThread}
+      testId="plugin-interaction-shell"
+    >
+      {() => (
+        <>
           <p className="mb-4 text-xs text-muted-foreground">
             {dismissal === "cancel"
               ? "Requested by "
@@ -182,16 +156,8 @@ export function PluginPendingInteractionComposer({
               </Button>
             </div>
           )}
-        </div>
-      </Activity>
-      {error ? (
-        <p
-          className="mx-4 mb-3 rounded-md border border-surface-destructive-border bg-surface-destructive px-2 py-1 text-xs text-destructive-text"
-          aria-live="polite"
-        >
-          {error}
-        </p>
-      ) : null}
-    </section>
+        </>
+      )}
+    </PendingInteractionShell>
   );
 }
