+++ apps/app/src/components/thread/pending-interactions/ThreadPendingInteractionBanner.tsx
@@ -1,12 +1,9 @@
 import {
-  useLayoutEffect,
-  useMemo,
-  useRef,
-  useState,
-  type KeyboardEvent,
-  type ReactNode,
-} from "react";
-import { NavLink } from "react-router-dom";
+  PendingInteractionShell,
+  type PendingInteractionSourceThread,
+  type PendingInteractionLayout,
+} from "./PendingInteractionShell";
+import { useMemo, useState, type ReactNode } from "react";
 import {
   assertNever,
   buildPendingInteractionApprovalResolution,
@@ -25,7 +22,6 @@ import {
   type PendingInteractionUserQuestionQuestion,
 } from "@bb/domain";
 import { Button } from "@bb/shared-ui/button";
-import { ExpandableLine } from "@/components/ui/expandable-line.js";
 import { Icon } from "@bb/shared-ui/icon";
 import { MarkdownPreview } from "@/components/ui/markdown-preview.js";
 import { getDetailScrollMaxHeightClass } from "@/components/ui/detail-scroll-size.js";
@@ -46,14 +42,9 @@ import { PluginCompactIconMask } from "@/components/plugin/PluginIcon";
 import { usePluginIconUrl } from "@/lib/plugin-logos";
 import { cn } from "@bb/shared-ui/lib/utils";
 
-interface ThreadPendingInteractionSourceThread {
-  href: string;
-  title: string;
-}
-
 interface ThreadPendingInteractionBannerProps {
   interaction: PendingInteraction;
-  sourceThread?: ThreadPendingInteractionSourceThread;
+  sourceThread?: PendingInteractionSourceThread;
   threadId: string;
 }
 
@@ -66,30 +57,16 @@ interface ApprovalPendingInteractionBannerProps {
   interaction: PendingInteraction;
   payload: ApprovalPendingInteractionPayload;
   subject: ApprovalBannerSubject;
-  sourceThread?: ThreadPendingInteractionSourceThread;
+  sourceThread?: PendingInteractionSourceThread;
   threadId: string;
 }
 
 interface UserQuestionPendingInteractionBannerProps {
   interaction: PendingInteraction;
   questions: readonly PendingInteractionUserQuestionQuestion[];
-  sourceThread?: ThreadPendingInteractionSourceThread;
+  sourceThread?: PendingInteractionSourceThread;
   threadId: string;
 }
-
-interface BannerShellProps {
-  label: string;
-  title?: string;
-  summary?: string | null;
-  initiallyExpanded: boolean;
-  errorMessage?: string | null;
-  footer?: (layout: BannerLayout) => ReactNode;
-  children?: (isExpanded: boolean) => ReactNode;
-  sourceThread?: ThreadPendingInteractionSourceThread;
-  testId: string;
-}
-
-type BannerLayout = "strip" | "card";
 
 interface ApprovalSubject {
   title: string;
@@ -157,15 +134,8 @@ function PendingInteractionBanner({
           data-testid="plugin-request-banner"
           data-request-kind={request.kind}
         >
-          {sourceThread ? (
-            <NavLink
-              to={sourceThread.href}
-              className="mb-1 block text-xs text-muted-foreground no-underline hover:underline"
-            >
-              From child thread: {sourceThread.title}
-            </NavLink>
-          ) : null}
           <PluginPendingInteractionComposer
+            sourceThread={sourceThread}
             interaction={interaction}
             request={{
               pluginId: request.pluginId,
@@ -185,7 +155,7 @@ function PendingInteractionBanner({
 interface PlanReviewRequestBannerProps {
   interaction: PendingInteraction;
   request: Extract<InteractionRequestView, { kind: "plan_review" }>;
-  sourceThread?: ThreadPendingInteractionSourceThread;
+  sourceThread?: PendingInteractionSourceThread;
   threadId: string;
 }
 
@@ -220,7 +190,7 @@ function PlanReviewRequestBanner({
   };
   const { plan, planFilePath } = request.review;
   return (
-    <BannerShell
+    <PendingInteractionShell
       label="Plan review"
       title={approval.reason ?? "Ready to code?"}
       summary={planFilePath ?? firstLine(plan)}
@@ -259,160 +229,7 @@ function PlanReviewRequestBanner({
           ) : null}
         </div>
       )}
-    </BannerShell>
-  );
-}
-
-function BannerShell({
-  label,
-  title,
-  summary,
-  initiallyExpanded,
-  errorMessage,
-  footer,
-  children,
-  sourceThread,
-  testId,
-}: BannerShellProps) {
-  const [isExpanded, setIsExpanded] = useState(initiallyExpanded);
-  const toggleRef = useRef<HTMLButtonElement>(null);
-  const shouldRestoreToggleFocusRef = useRef(false);
-  const collapsesOnEscape = !initiallyExpanded;
-  useLayoutEffect(() => {
-    if (!shouldRestoreToggleFocusRef.current) return;
-    shouldRestoreToggleFocusRef.current = false;
-    toggleRef.current?.focus();
-  }, [isExpanded]);
-  const handleKeyDown = (event: KeyboardEvent<HTMLElement>) => {
-    if (
-      event.key === "Escape" &&
-      isExpanded &&
-      collapsesOnEscape &&
-      !event.defaultPrevented
-    ) {
-      event.preventDefault();
-      event.stopPropagation();
-      shouldRestoreToggleFocusRef.current = true;
-      setIsExpanded(false);
-    }
-  };
-  const toggle = (
-    <button
-      ref={toggleRef}
-      type="button"
-      aria-expanded={isExpanded}
-      aria-label={isExpanded ? "Hide details" : "Show details"}
-      onClick={(event) => {
-        shouldRestoreToggleFocusRef.current =
-          document.activeElement === event.currentTarget;
-        setIsExpanded((value) => !value);
-      }}
-      className="flex size-7 shrink-0 items-center justify-center rounded-md text-muted-foreground hover:bg-state-hover hover:text-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring"
-    >
-      <Icon
-        name="ChevronDown"
-        className={cn(
-          "size-3.5 transition-transform duration-200",
-          isExpanded ? "rotate-180" : undefined,
-        )}
-      />
-    </button>
-  );
-  const errorNode = errorMessage ? (
-    <div className="mx-3 mb-3 rounded-md border border-surface-destructive-border bg-surface-destructive px-2 py-1 text-xs text-destructive-text">
-      {errorMessage}
-    </div>
-  ) : null;
-  const sourceThreadLink = sourceThread ? (
-    <NavLink
-      to={sourceThread.href}
-      title={sourceThread.title}
-      className="min-w-24 shrink-[3] truncate text-xs text-subtle-foreground no-underline hover:underline"
-    >
-      From {sourceThread.title}
-    </NavLink>
-  ) : null;
-
-  return (
-    <section
-      aria-label={label}
-      data-testid={testId}
-      data-expanded={isExpanded ? "" : undefined}
-      onKeyDown={handleKeyDown}
-      className="@container mb-2 min-w-0 max-w-full rounded-lg border border-attention/40 bg-surface-raised-solid text-xs text-muted-foreground ring-[3px] ring-surface-attention"
-    >
-      {isExpanded ? (
-        <div className="flex items-center gap-2 border-b border-border-hairline py-1.5 pl-3 pr-1.5">
-          <AttentionDot />
-          <span className="shrink-0 text-sm font-semibold text-foreground">
-            {label}
-          </span>
-          {sourceThreadLink}
-          <span className="flex-1" />
-          {toggle}
-        </div>
-      ) : (
-        <div className="flex min-h-9 items-center gap-2 py-1 pl-3 pr-1.5">
-          <div className="flex min-w-0 flex-1 items-center gap-2">
-            <AttentionDot />
-            <div className="flex min-w-0 flex-1 items-center gap-2">
-              <span
-                className="min-w-0 truncate text-sm font-medium text-foreground @2xl:shrink"
-                title={title ?? label}
-              >
-                {title ?? label}
-              </span>
-              {summary ? (
-                <span
-                  className="hidden min-w-0 truncate font-mono text-xs text-muted-foreground @2xl:block @2xl:shrink-[2]"
-                  title={summary}
-                >
-                  {summary}
-                </span>
-              ) : null}
-              <div className="hidden min-w-0 @2xl:contents">
-                {sourceThreadLink}
-              </div>
-            </div>
-          </div>
-          <div className="order-2 shrink-0 @2xl:order-3">{toggle}</div>
-          {footer ? (
-            <div className="order-2 hidden shrink-0 items-center gap-1.5 @2xl:flex">
-              {footer("strip")}
-            </div>
-          ) : null}
-        </div>
-      )}
-      <div hidden={!isExpanded} className="px-3 pb-3 pt-2.5">
-        {title ? (
-          <h3 className="min-w-0 text-sm font-medium text-foreground">
-            <ExpandableLine fullText={title} collapsedClassName="line-clamp-2">
-              {title}
-            </ExpandableLine>
-          </h3>
-        ) : null}
-        {children ? (
-          <div className={title ? "mt-2" : undefined}>
-            {children(isExpanded)}
-          </div>
-        ) : null}
-        {footer ? (
-          <div className="mt-3 flex flex-wrap items-center gap-2">
-            {footer("card")}
-          </div>
-        ) : null}
-      </div>
-      {errorNode}
-    </section>
-  );
-}
-
-function AttentionDot() {
-  return (
-    <span
-      aria-hidden="true"
-      className="size-2 shrink-0 rounded-full bg-attention ring-[3px] ring-surface-attention"
-    />
+    </PendingInteractionShell>
   );
 }
 
@@ -474,7 +291,7 @@ function ApprovalPendingInteractionBanner({
   };
 
   return (
-    <BannerShell
+    <PendingInteractionShell
       label="Approval needed"
       title={view.title}
       summary={view.summary}
@@ -494,7 +311,7 @@ function ApprovalPendingInteractionBanner({
       )}
     >
       {() => view.body}
-    </BannerShell>
+    </PendingInteractionShell>
   );
 }
 
@@ -507,7 +324,7 @@ function ThreadUserQuestionPendingInteractionBanner({
   const isResolving = interaction.status === "resolving";
 
   return (
-    <BannerShell
+    <PendingInteractionShell
       label={
         questions.length === 1 ? "Question" : `${questions.length} questions`
       }
@@ -525,14 +342,14 @@ function ThreadUserQuestionPendingInteractionBanner({
           threadId={threadId}
         />
       )}
-    </BannerShell>
+    </PendingInteractionShell>
   );
 }
 
 interface ApprovalDecisionButtonsProps {
   decisions: readonly PendingInteractionApprovalDecision[];
   disabled: boolean;
-  layout: BannerLayout;
+  layout: PendingInteractionLayout;
   loadingDecision: PendingInteractionApprovalDecision | null;
   onDecide: (decision: PendingInteractionApprovalDecision) => void;
   subjectKind: PendingInteractionApprovalSubject["kind"];
@@ -570,7 +387,7 @@ function ApprovalDecisionButtons({
 
 interface ApprovalDecisionButtonProps {
   className?: string;
-  layout: BannerLayout;
+  layout: PendingInteractionLayout;
   decision: PendingInteractionApprovalDecision;
   disabled: boolean;
   isLoading: boolean;
